+++ backend/mysell/src/main/resources/db/migration/V1__create_tables.sql
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS units_of_measure (
 );
 CREATE TABLE IF NOT EXISTS products_units_of_measure (
     products_units_of_measure_id SERIAL PRIMARY KEY,
-    quantity INT DEFAULT 0 NOT NULL,
+    quantity DOUBLE PRECISION DEFAULT 0 NOT NULL,
     unit_of_measure_id BIGINT DEFAULT 1 NOT NULL
 );
 CREATE TABLE IF NOT EXISTS products (
